// src/shared/errors/app-error.enum.ts
export enum ErrorEnum {
  BAD_REQUEST = 'BAD_REQUEST',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  URL_NOT_FOUND = 'URL_NOT_FOUND',
  DATABASE_ERROR = 'DATABASE_ERROR',
  ALREADY_EXISTS = 'ALREADY_EXISTS',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  DAILY_LIMIT_REACHED = 'DAILY_LIMIT_REACHED',
  CANNOT_FOLLOW_SELF = 'CANNOT_FOLLOW_SELF',
  CANNOT_UNFOLLOW_SELF = 'CANNOT_UNFOLLOW_SELF',

}